+++ src/auth/utils/createNonce.ts
@@ -1,7 +1,8 @@
 import {
   WalletInterface,
   WalletCounterparty,
-  Base64String
+  Base64String,
+  OriginatorDomainNameStringUnder250Bytes
 } from '../../wallet/Wallet.interfaces.js'
 import * as Utils from '../../primitives/utils.js'
 import Random from '../../primitives/Random.js'
@@ -14,7 +15,8 @@ import Random from '../../primitives/Random.js'
  */
 export async function createNonce(
   wallet: WalletInterface,
-  counterparty: WalletCounterparty = 'self'
+  counterparty: WalletCounterparty = 'self',
+  originator?: OriginatorDomainNameStringUnder250Bytes
 ): Promise<Base64String> {
   // Generate 16 random bytes for the first half of the data
   const firstHalf = Random(16)
@@ -24,7 +26,7 @@ export async function createNonce(
     keyID: Utils.toUTF8(firstHalf),
     data: firstHalf,
     counterparty
-  })
+  }, originator)
   // Concatenate firstHalf and secondHalf as the nonce bytes
   const nonceBytes = [...firstHalf, ...hmac]
   return Utils.toBase64(nonceBytes)